// src/domain/enterprise/entities/account/enums/permissions-type.ts
export enum PermissionType {
    DELIVER_CREATE = "DELIVER_CREATE",
    DELIVER_UPDATE = "DELIVER_UPDATE",
    DELIVER_DELETE = "DELIVER_DELETE",

    DELIVERER_CREATE = "DELIVERER_CREATE",
    DELIVERER_UPDATE = "DELIVERER_UPDATE",
    DELIVERER_DELETE = "DELIVERER_DELETE",

    ADMIN_VIEW = "ADMIN_VIEW",
    ADMIN_CREATE = "ADMIN_CREATE",
    ADMIN_UPDATE = "ADMIN_UPDATE",
    ADMIN_DELETE = "ADMIN_DELETE",

    RECIPIENT_DELETE = "RECIPIENT_DELETE",
    RECIPIENT_CREATE = "RECIPIENT_CREATE",
    RECIPIENT_UPDATE = "RECIPIENT_UPDATE",


    VIEW = "VIEW",
    CREATE = "CREATE",
    UPDATE = "UPDATE",
    DELETE = "DELETE",

}